beacon-chain/sync/rpc_hello_test.go
package sync

import (
	"context"
	"testing"

	p2ptest "github.com/prysmaticlabs/prysm/beacon-chain/p2p/testing"
	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
)

func TestHelloRPCHandler_Disconnects_OnForkVersionMismatch(t *testing.T) {
	p1 := p2ptest.NewTestP2P(t)
	p2 := p2ptest.NewTestP2P(t)
	p1.Connect(p2)
	if len(p1.Host.Network().Peers()) != 1 {
		t.Error("Expected peers to be connected")
	}

	r := RegularSync{p2p: p1}

	stream, err := p1.Swarm.NewStream(context.Background(), p2.Host.ID())
	if err != nil {
		t.Fatal(err)
	}

	err = r.helloRPCHandler(context.Background(), &pb.Hello{ForkVersion: []byte("ff")}, stream)
	if err != errWrongForkVersion {
		t.Errorf("Expected error %v, got %v", errWrongForkVersion, err)
	}

	if len(p1.Host.Network().Peers()) != 0 {
		t.Error("handler did not disconnect peer")
	}
}
